Update Clear all completed items functionality
